requirements.txt
google-adk
a2a-sdk
fastmcp
mcp
langgraph
langchain-core
langchain-google-genai
google-generativeai
httpx
uvicorn
nest-asyncio
python-dotenv

